+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed deprecated `Config.linesOfCodeWarningLevel` use `Config.linesOfExecutableCodeWarningLevel` instead
 - Removed deprecated `MetricsAnalysisRecorder.startRecordFile` and `MetricsAnalysisRecorder.endRecordFile` use `MetricsRecordsStore.recordFile` instead
 - Add static code anti-patterns long-parameter-list
+- Add static code diagnostic no-equal-arguments
 - Set min SDK version to >=2.8.0.
 
 # 1.10.0